refactor: Header component'inin yeri degistirildi.
